pyproject.toml
[project]
name = "landly"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependency-groups.dev = [
    "flet[all]==0.28.3",
]
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
  "fastapi>=0.119.0",
  "flet[all]==0.28.3",
  "flet-map>=0.1.0",
  "mkdocs>=1.6.1",
  "postgis>=1.0.4",
  "postgres>=4.0",
  "sqlalchemy>=2.0.44",
  "uvicorn>=0.38.0",
  "requests>=2.32.5",
  "python-dotenv>=1.2.1",
  "geoalchemy2>=0.17.1",
  "pydantic[email]>=2.12.5",
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.mycompany"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "landly"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2025 by Flet"

[tool.flet.app]
path = "src"


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
flet = {extras = ["all"], version = "0.28.3"}
